# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "iLaw API",
    "version": "1.0.0",
    "description": "아이법 백엔드 API 문서"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGci..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGci..."
          },
          "profileCompleted": {
            "type": "boolean",
            "description": "온보딩(추가 정보 입력) 완료 여부. false인 경우 PATCH /auth/profile 호출 필요",
            "example": false
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmorkvmmx0000139yevgthkpz"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "user@example.com"
          },
          "nickname": {
            "type": "string",
            "nullable": true,
            "example": "홍길동"
          },
          "region": {
            "type": "string",
            "nullable": true,
            "example": "서울특별시"
          },
          "birthYear": {
            "type": "integer",
            "nullable": true,
            "example": 1995
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ],
            "nullable": true
          },
          "provider": {
            "type": "string",
            "enum": [
              "kakao",
              "google"
            ],
            "example": "kakao"
          },
          "profileCompleted": {
            "type": "boolean",
            "example": true
          },
          "agreedMarketing": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ManualCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "아동학대"
          },
          "slug": {
            "type": "string",
            "example": "child-abuse"
          },
          "order": {
            "type": "integer"
          }
        }
      },
      "ManualArticleSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "question": {
            "type": "string",
            "example": "친권은 무엇이고 누가 친권자가 되나요?"
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "type": "integer"
          }
        }
      },
      "ManualArticleDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "question": {
            "type": "string"
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "description": "마크다운 형식의 본문"
          },
          "category": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              }
            }
          }
        }
      },
      "Agency": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "region": {
            "type": "string",
            "example": "서울"
          },
          "name": {
            "type": "string",
            "example": "금융감독원"
          },
          "role": {
            "type": "string",
            "example": "금융사기·불법대출 상담"
          },
          "contact": {
            "type": "string",
            "example": "1332"
          }
        }
      }
    }
  },
  "paths": {
    "/auth/kakao": {
      "get": {
        "summary": "카카오 로그인",
        "description": "서명된 state와 브라우저 트랜잭션 쿠키를 발급하고 카카오 로그인 페이지로 리다이렉트합니다.",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "web",
                "local"
              ]
            },
            "description": "고정 리다이렉트 대상. local은 비운영 환경에서만 허용됩니다."
          }
        ],
        "responses": {
          "302": {
            "description": "카카오 로그인 페이지로 리다이렉트"
          },
          "400": {
            "description": "허용되지 않은 target 또는 redirectUri 입력"
          }
        }
      }
    },
    "/auth/kakao/callback": {
      "get": {
        "summary": "카카오 로그인 콜백",
        "description": "state와 브라우저 트랜잭션을 검증한 뒤 고정된 프론트엔드 주소로 일회용 code를 전달합니다. JWT는 URL에 포함하지 않습니다.",
        "tags": [
          "Auth"
        ],
        "responses": {
          "303": {
            "description": "로그인 성공 후 ?code=...가 포함된 고정 프론트엔드 주소로 리다이렉트"
          },
          "400": {
            "description": "유효하지 않거나 만료되었거나 이미 사용된 state"
          }
        }
      }
    },
    "/auth/google": {
      "get": {
        "summary": "구글 로그인",
        "description": "서명된 state와 브라우저 트랜잭션 쿠키를 발급하고 구글 로그인 페이지로 리다이렉트합니다.",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "web",
                "local"
              ]
            },
            "description": "고정 리다이렉트 대상. local은 비운영 환경에서만 허용됩니다."
          }
        ],
        "responses": {
          "302": {
            "description": "구글 로그인 페이지로 리다이렉트"
          },
          "400": {
            "description": "허용되지 않은 target 또는 redirectUri 입력"
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "summary": "구글 로그인 콜백",
        "description": "state와 브라우저 트랜잭션을 검증한 뒤 고정된 프론트엔드 주소로 일회용 code를 전달합니다. JWT는 URL에 포함하지 않습니다.",
        "tags": [
          "Auth"
        ],
        "responses": {
          "303": {
            "description": "로그인 성공 후 ?code=...가 포함된 고정 프론트엔드 주소로 리다이렉트"
          },
          "400": {
            "description": "유효하지 않거나 만료되었거나 이미 사용된 state"
          }
        }
      }
    },
    "/auth/exchange": {
      "post": {
        "summary": "OAuth 일회용 코드 교환",
        "description": "OAuth 콜백으로 받은 일회용 code를 iLaw JWT로 교환합니다. code는 1분간 유효하며 한 번만 사용할 수 있습니다.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "교환 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "code 누락"
          },
          "401": {
            "description": "유효하지 않거나 만료되었거나 이미 사용된 code"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Access Token 재발급",
        "description": "Refresh Token으로 새로운 Access Token을 발급받습니다.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGci..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "토큰 재발급 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "유효하지 않은 Refresh Token"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "로그아웃",
        "description": "Refresh Token을 무효화합니다. Access Token이 필요합니다.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "로그아웃 성공"
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "내 정보 조회",
        "description": "로그인한 유저의 정보를 반환합니다. Access Token이 필요합니다.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "유저 정보 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    },
    "/auth/profile": {
      "patch": {
        "summary": "온보딩 - 추가 정보 입력 및 약관 동의",
        "description": "소셜 로그인 후 닉네임, 지역, 출생연도, 성별, 약관 동의를 저장하고 profileCompleted를 true로 변경합니다.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nickname",
                  "region",
                  "birthYear",
                  "gender",
                  "agreedTermsOfService",
                  "agreedPrivacyPolicy",
                  "agreedAge14"
                ],
                "properties": {
                  "nickname": {
                    "type": "string",
                    "example": "홍길동"
                  },
                  "region": {
                    "type": "string",
                    "example": "서울특별시"
                  },
                  "birthYear": {
                    "type": "integer",
                    "example": 1995
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "male",
                      "female",
                      "other"
                    ]
                  },
                  "agreedTermsOfService": {
                    "type": "boolean"
                  },
                  "agreedPrivacyPolicy": {
                    "type": "boolean"
                  },
                  "agreedAge14": {
                    "type": "boolean"
                  },
                  "agreedMarketing": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "프로필 업데이트 성공"
          },
          "400": {
            "description": "필수 필드 누락 또는 유효하지 않은 값"
          },
          "409": {
            "description": "닉네임 중복"
          }
        }
      }
    },
    "/auth/kakao/token": {
      "post": {
        "summary": "카카오 SDK 로그인",
        "description": "카카오 SDK로 받은 accessToken으로 JWT를 발급합니다.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accessToken"
                ],
                "properties": {
                  "accessToken": {
                    "type": "string",
                    "example": "eyJhbGci..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "유효하지 않은 토큰"
          }
        }
      }
    },
    "/auth/google/token": {
      "post": {
        "summary": "구글 SDK 로그인",
        "description": "Google Sign-In SDK로 받은 idToken으로 JWT를 발급합니다.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idToken"
                ],
                "properties": {
                  "idToken": {
                    "type": "string",
                    "example": "eyJhbGci..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "유효하지 않은 토큰"
          }
        }
      }
    },
    "/manual/categories": {
      "get": {
        "summary": "매뉴얼 카테고리 목록",
        "tags": [
          "Manual"
        ],
        "responses": {
          "200": {
            "description": "카테고리 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualCategory"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manual/categories/{slug}/articles": {
      "get": {
        "summary": "카테고리별 아티클 목록",
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "child-abuse"
          }
        ],
        "responses": {
          "200": {
            "description": "아티클 목록 (본문 제외)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualArticleSummary"
                  }
                }
              }
            }
          },
          "404": {
            "description": "카테고리 없음"
          }
        }
      }
    },
    "/manual/articles/{id}": {
      "get": {
        "summary": "아티클 상세 조회",
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "아티클 상세 (본문 포함)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualArticleDetail"
                }
              }
            }
          },
          "404": {
            "description": "아티클 없음"
          }
        }
      }
    },
    "/manual/categories/{slug}/agencies": {
      "get": {
        "summary": "카테고리별 기관 연락망",
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "child-abuse"
          },
          {
            "in": "query",
            "name": "region",
            "schema": {
              "type": "string"
            },
            "example": "서울",
            "description": "지역 필터 (생략 시 전체)"
          }
        ],
        "responses": {
          "200": {
            "description": "기관 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agency"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}